Changed the vertex bone assignments to unsigned int rather than unsigned short, ie 32 bit. Reimplemented Mesh::compileBoneAssignments

+++ OgreMain/include/OgreSubMesh.h
@@ -115,7 +115,7 @@ namespace Ogre {
         void clearBoneAssignments(void);
 
         /// Multimap of verex bone assignments (orders by vertex index)
-        typedef std::multimap<unsigned short, VertexBoneAssignment> VertexBoneAssignmentList;
+        typedef std::multimap<size_t, VertexBoneAssignment> VertexBoneAssignmentList;
         typedef MapIterator<VertexBoneAssignmentList> BoneAssignmentIterator;
 
         /** Gets an iterator for access all bone assignments.
